Fix/update context logic (#159)
* Update contex logic by using one file

* Use function that returns context

* Move context values into an object

* Divide context values into getters and setters
