+++ src/components/MobilityPlatform/Boating/BoatParking/BoatParking.js
@@ -1,8 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useMap } from 'react-leaflet';
-import MobilityPlatformContext from '../../../../context/MobilityPlatformContext';
+import { useMobilityPlatformContext } from '../../../../context/MobilityPlatformContext';
 import { useAccessibleMap } from '../../../../redux/selectors/settings';
 import {
   isDataValid, fitPolygonsToBounds, blueOptionsBase, whiteOptionsBase,
@@ -18,7 +18,7 @@ import TextContent from '../../TextContent';
 const BoatParking = () => {
   const [boatParkingData, setBoatParkingData] = useState([]);
 
-  const { openMobilityPlatform, showBoatParking } = useContext(MobilityPlatformContext);
+  const { openMobilityPlatform, showBoatParking } = useMobilityPlatformContext();
 
   const useContrast = useSelector(useAccessibleMap);
