+++ src/components/MobilityPlatform/Boating/GuestHarbour/GuestHarbour.js
@@ -1,8 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useMap } from 'react-leaflet';
 import { useSelector } from 'react-redux';
-import MobilityPlatformContext from '../../../../context/MobilityPlatformContext';
+import { useMobilityPlatformContext } from '../../../../context/MobilityPlatformContext';
 import { useAccessibleMap } from '../../../../redux/selectors/settings';
 import { fetchMobilityMapPolygonData } from '../../mobilityPlatformRequests/mobilityPlatformRequests';
 import {
@@ -18,7 +18,7 @@ import TextContent from '../../TextContent';
 const GuestHarbour = () => {
   const [guestHarbourData, setGuestHarbourData] = useState([]);
 
-  const { openMobilityPlatform, showGuestHarbour } = useContext(MobilityPlatformContext);
+  const { openMobilityPlatform, showGuestHarbour } = useMobilityPlatformContext();
 
   const useContrast = useSelector(useAccessibleMap);
