+++ src/components/MobilityPlatform/ChargerStationMarkers/ChargerStationMarkers.js
@@ -1,10 +1,10 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useMap } from 'react-leaflet';
 import chargerIcon from 'servicemap-ui-turku/assets/icons/icons-icon_charging_station.svg';
 import chargerIconBw from 'servicemap-ui-turku/assets/icons/contrast/icons-icon_charging_station-bw.svg';
-import MobilityPlatformContext from '../../../context/MobilityPlatformContext';
+import { useMobilityPlatformContext } from '../../../context/MobilityPlatformContext';
 import { useAccessibleMap } from '../../../redux/selectors/settings';
 import { fetchMobilityMapData } from '../mobilityPlatformRequests/mobilityPlatformRequests';
 import { createIcon, isDataValid, fitToMapBounds } from '../utils/utils';
@@ -14,7 +14,7 @@ import ChargerStationContent from './components/ChargerStationContent';
 const ChargerStationMarkers = () => {
   const [chargerStations, setChargerStations] = useState([]);
 
-  const { openMobilityPlatform, showChargingStations } = useContext(MobilityPlatformContext);
+  const { openMobilityPlatform, showChargingStations } = useMobilityPlatformContext();
 
   const map = useMap();
