+++ src/components/MobilityPlatform/FitnessTrails/FitnessTrails.js
@@ -1,11 +1,11 @@
-import React, { useContext } from 'react';
-import MobilityPlatformContext from '../../../context/MobilityPlatformContext';
+import React from 'react';
+import { useMobilityPlatformContext } from '../../../context/MobilityPlatformContext';
 import Trails from '../TrailsComponent';
 
 /* Show selected fitness trail on the map */
 
 const FitnessTrails = () => {
-  const { showFitnessTrails, fitnessTrailsObj } = useContext(MobilityPlatformContext);
+  const { showFitnessTrails, fitnessTrailsObj } = useMobilityPlatformContext();
 
   const tealColor = 'rgba(102, 102, 153, 255)';
   const dashPattern = '10 2 7';
