+++ src/components/MobilityPlatform/LoadingPlaces/LoadingPlaces.js
@@ -1,9 +1,10 @@
-import React, { useContext, useEffect, useState } from 'react';
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect, useState } from 'react';
 import { useMap } from 'react-leaflet';
 import { useSelector } from 'react-redux';
 import loadingPlaceIcon from 'servicemap-ui-turku/assets/icons/icons-icon_loading_place.svg';
 import loadingPlaceIconBw from 'servicemap-ui-turku/assets/icons/contrast/icons-icon_loading_place-bw.svg';
-import MobilityPlatformContext from '../../../context/MobilityPlatformContext';
+import { useMobilityPlatformContext } from '../../../context/MobilityPlatformContext';
 import { useAccessibleMap } from '../../../redux/selectors/settings';
 import { fetchMobilityMapPolygonData } from '../mobilityPlatformRequests/mobilityPlatformRequests';
 import { isDataValid, fitPolygonsToBounds, createIcon } from '../utils/utils';
@@ -12,7 +13,7 @@ import LoadingPlacesContent from './components/LoadingPlacesContent';
 const LoadingPlaces = () => {
   const [loadingPlaces, setLoadingPlaces] = useState([]);
 
-  const { openMobilityPlatform, showLoadingPlaces } = useContext(MobilityPlatformContext);
+  const { openMobilityPlatform, showLoadingPlaces } = useMobilityPlatformContext();
 
   const map = useMap();
