+++ src/components/MobilityPlatform/MarkedTrails/MarkedTrails.js
@@ -1,11 +1,11 @@
-import React, { useContext } from 'react';
-import MobilityPlatformContext from '../../../context/MobilityPlatformContext';
+import React from 'react';
+import { useMobilityPlatformContext } from '../../../context/MobilityPlatformContext';
 import Trails from '../TrailsComponent';
 
 /* Show marked trails which are part of Paavo trails on the map */
 
 const MarkedTrails = () => {
-  const { showMarkedTrails, markedTrailsObj } = useContext(MobilityPlatformContext);
+  const { showMarkedTrails, markedTrailsObj } = useMobilityPlatformContext();
 
   const brownColor = 'rgba(117, 44, 23, 255)';
   const dashPattern = '2 9 9 9';
