+++ src/components/MobilityPlatform/NatureTrails/NatureTrails.js
@@ -1,11 +1,11 @@
-import React, { useContext } from 'react';
-import MobilityPlatformContext from '../../../context/MobilityPlatformContext';
+import React from 'react';
+import { useMobilityPlatformContext } from '../../../context/MobilityPlatformContext';
 import Trails from '../TrailsComponent';
 
 /* Show marked trails which are part of Paavo trails on the map */
 
 const NatureTrails = () => {
-  const { showNatureTrails, natureTrailsObj } = useContext(MobilityPlatformContext);
+  const { showNatureTrails, natureTrailsObj } = useMobilityPlatformContext();
 
   const blueColor = 'rgba(0, 0, 153, 255)';
   const dashPattern = '12 9';
