+++ src/components/MobilityPlatform/Scooters/components/NoParking/NoParking.js
@@ -1,8 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useMap } from 'react-leaflet';
 import { useSelector } from 'react-redux';
-import MobilityPlatformContext from '../../../../../context/MobilityPlatformContext';
+import { useMobilityPlatformContext } from '../../../../../context/MobilityPlatformContext';
 import { fetchMobilityMapPolygonData } from '../../../mobilityPlatformRequests/mobilityPlatformRequests';
 import { isDataValid, whiteOptionsBase, redOptionsBase } from '../../../utils/utils';
 import { useAccessibleMap } from '../../../../../redux/selectors/settings';
@@ -16,7 +16,7 @@ import TextContent from '../../../TextContent';
 const NoParking = () => {
   const [noParkingData, setNoParkingData] = useState([]);
 
-  const { openMobilityPlatform, showScooterNoParking } = useContext(MobilityPlatformContext);
+  const { openMobilityPlatform, showScooterNoParking } = useMobilityPlatformContext();
 
   useEffect(() => {
     if (openMobilityPlatform) {
